Added type annotations to the introspection code which instantiates the tested lexer/parser/treeparser.
This should cause the code to fail earlier when it gets into an invalid state, and perhaps even give a better error message.

On the other hand though, if we want to allow testing of Parsers, Lexers, etc which have the same interface as antlr parsers but which don't subclass antlr's Parser classes, then this patch isn't worthwhile.